* Updated Eden Quests 71-85 to its latest version. (translation: Muad_Dib) * Updated all 1st class training quests, mostly experience changes. (credits: Muad_Dib, partial revert of r17210) * Added Pre-Renewal support to 1st class training quests. (bugreport:7808) * Corrected level 10 AL_HEAL lines in 'npcskill' to use MVP heal. * Corrected EXP checks in 'checkre' to use RENEWAL_EXP instead of RENEWAL. * Fixed Hunter skill quest not accepting Rangers. * Fixed a bug in Hall of Abyss. (bugreport:7815) * Fix unable to move when cloaking in levels less than 3. (credits: Akinari, bugreport:7810) * Idle timer now resets on continuous attacks, i.e. /noctrl or ctrl+click. (Hercules 19f0fcd) * Idle timer is now initialized upon login, fixing a bug with the 'checkidle' command. (credits: Akinari)
git-svn-id: https://svn.code.sf.net/p/rathena/svn/trunk@17409 54d463be-8e91-2dee-dedb-b68131a5f0ec

+++ src/map/battle.c
@@ -4356,9 +4356,9 @@ static struct Damage battle_calc_weapon_attack(struct block_list *src, struct bl
 #ifdef RENEWAL
 	if(is_attack_critical(wd, src, target, skill_id, skill_lv, false)) {
 		if(sd) // check for player so we don't crash out, monsters don't have bonus crit rates [helvetica]
-			wd.damage = (int32)floor(wd.damage * 1.4 * ((100 + sd->bonus.crit_atk_rate) / 100));
+			wd.damage = (int32)floor((double)wd.damage * 1.4 * ((100 + sd->bonus.crit_atk_rate) / 100));
 		else
-			wd.damage = (int32)floor(wd.damage * 1.4);
+			wd.damage = (int32)floor((double)wd.damage * 1.4);
 	}
 #endif
